refactor: set directory structure for tests
