+++ src/web/src/__tests__/integration/infrastructure.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
 import { screen } from "@testing-library/react";
-import { render } from "./test-utils";
+import { render } from "../test-utils";
 
 function TestComponent() {
   return <div>Hello Testing World!</div>;
